==> 7-feature-selection.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('filter_variable_list.csv')
df.head()

df = df[df.columns[1:]]

df.dtypes

#Then remove the first 2 weeks and the last 4 months (OOT)
len(df)

df['random'] = np.random.normal(0, 1, size=(1, len(df)))[0]
df['random']

from scipy import stats as sps

goods = df[df.Fraud == 0]
bads = df[df.Fraud == 1]

KSFDR = pd.DataFrame({})

i = 0
for column in df:
    KSFDR.loc[i, 'ks'] = sps.ks_2samp(goods[column],bads[column])[0]
    i = i+1
    
KSFDR.index = df.columns

KSFDR

topRows = int(round(len(df)*0.03))
j = 0
numbads = len(bads)

for column in df:
    temp = df.sort_values(column,ascending=False)
    temp1 = temp.head(topRows)
    temp2 = temp.tail(topRows)
    needed1 = temp1.loc[:,'Fraud']
    needed2 = temp2.loc[:,'Fraud']
    FDR1 = sum(needed1)/numbads
    FDR2 = sum(needed2)/numbads
    FDRate = np.maximum(FDR1,FDR2)
    KSFDR.loc[column, 'FDR'] = FDRate
    j = j +1


KSFDR

KSFDR['rank_ks'] = KSFDR['ks'].rank(ascending = True)
KSFDR['rank_FDR'] = KSFDR['FDR'].rank(ascending = True)
KSFDR['average_rank'] = (KSFDR['rank_ks'] + KSFDR['rank_FDR']) / 2
KSFDR.sort_values(by=['average_rank'], ascending=False, inplace=True)


KSFDR.loc['random']

KSFDR.reset_index().rename(columns={'index':'field'}).to_csv('finish-filter-list.csv', index = False)

#KSFDR.to_csv('finish-filter-list.csv')

def fdr_prob(y, y_prob, cutoff=0.03):
    if len(y_prob.shape) != 1:    # sometimes the proba list can contain many columns, one for each category
        y_prob = y_prob[:, -1:]   # only the last one (fraud_label==1) is used here.
    num_fraud = len(y[y == 1])    # count the total nunber of frauds   
# sort the proba list from high to low while retain the true (not predicted) fraud label
    sorted_prob = np.asarray(sorted(zip(y_prob, y), key=lambda x: x[0], reverse=True))
    cutoff_bin = sorted_prob[0:int(len(y) * cutoff), 1:]  # 3% cutoff
# return the FDR score (#fraud_in_cutoff / #total_fraud)
    return len(cutoff_bin[cutoff_bin == 1]) / num_fraud 


def fdr(classifier, x, y, cutoff=0.03):
# Calculates FDR score for the given classifier on dataset x and y with cutoff value
# get the probability list from the given classifier
    return fdr_prob(y, classifier.predict_proba(x), cutoff)

top_80 = KSFDR.reset_index().iloc[:81]['index'].to_list()[1:]

data = df[top_80]

Y = df['Fraud'].values

pip install mlxtend  

#wrapper
from mlxtend.feature_selection import SequentialFeatureSelector as SFS
from sklearn.ensemble import RandomForestClassifier

nfeatures = len(data.columns)
clf = RandomForestClassifier(n_estimators=5, n_jobs=-1)
sfs = SFS(clf,k_features=nfeatures,forward=True,verbose=2,scoring=fdr,cv=2,n_jobs=-1)
# sfs = SFS(clf,k_features=1,forward=False,verbose=2,scoring=fdr,cv=2,n_jobs=-1)
sfs.fit(data,Y)


vars_FS = pd.DataFrame.from_dict(sfs.get_metric_dict()).T
vars_30 = list(vars_FS.loc[30, 'feature_names'])
vars_FS.to_csv('vars_FS.csv')

importance = []
for i in range(30):
    for var in vars_FS.loc[i+1,'feature_names']:
        if var not in importance:
            importance.append(var)
            


top_30 = pd.DataFrame(importance, columns=['field'])
top_30.to_csv('top_30_wrapper.csv', index = False)
