==> 8-benford-low-card.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('all_df_var.csv')

df = df[df.columns[1:]]
df.head()

df = df[['Recnum', 'Cardnum', 'Merchnum', 'Merch description', 'Amount']]

idx = []
for i in range(len(df)):
    if 'FEDEX' in df.loc[i, 'Merch description']:
        idx.append(df.loc[i, 'Recnum'])

df = df.set_index(df.Recnum)

df = df.drop(labels=idx, axis=0)

df.loc[:,'Amount'] = df.loc[:,'Amount'].astype('str')

df.loc[:,'Amount_digit'] = [df.loc[i, 'Amount'][0] for i in df.index]

df_card = df[(df['Amount_digit'] == '1') | (df['Amount_digit'] == '2')]

df_low = df_card.groupby(['Cardnum'])['Recnum'].count().reset_index().rename(columns = {'Recnum':'n_low'})
df_card_sum = df.groupby('Cardnum')['Recnum'].count().reset_index().rename(columns = {'Recnum':'n'})
df_card_sum = df_card_sum.merge(df_low, how='left', on='Cardnum')
df_card_sum['n_low'] = df_card_sum['n_low'].fillna(0)
df_card_sum['n_high'] = df_card_sum['n'] - df_card_sum['n_low']

df_card_sum['n_low'] = df_card_sum['n_low'].replace(0,1)
df_card_sum['n_high'] = df_card_sum['n_high'].replace(0,1)

df_merc = df[(df['Amount_digit'] == '1') | (df['Amount_digit'] == '2')]
df_low_m = df_merc.groupby(['Merchnum'])['Recnum'].count().reset_index().rename(columns = {'Recnum':'n_low'})
df_sum_m = df.groupby(['Merchnum'])['Recnum'].count().reset_index().rename(columns = {'Recnum':'n'})
df_sum_m = df_sum_m.merge(df_low_m, how='left', on='Merchnum')

df_sum_m['n_low'] = df_sum_m['n_low'].fillna(0)
df_sum_m['n_high'] = df_sum_m['n'] - df_sum_m['n_low']
df_sum_m['n_low'] = df_sum_m['n_low'].replace(0,1)
df_sum_m['n_high'] = df_sum_m['n_high'].replace(0,1)


df_card_sum['R'] = (1.096 * df_card_sum['n_low'])/df_card_sum['n_high']
df_sum_m['R'] = (1.096 * df_sum_m['n_low'])/df_sum_m['n_high']

df_card_sum['1/R'] = 1/df_card_sum['R']
df_sum_m['1/R'] = 1/df_sum_m['R']

df_card_sum.loc[:,'U'] = [max(df_card_sum.loc[i,'R'], df_card_sum.loc[i,'1/R']) for i in df_card_sum.index]
df_sum_m.loc[:,'U'] = [max(df_sum_m.loc[i,'R'], df_sum_m.loc[i,'1/R']) for i in df_sum_m.index]



df_card_sum['t'] = (df_card_sum['n']-15)/3
df_sum_m['t'] = (df_sum_m['n']-15)/3

df_card_sum['U*']  = 1 + ((df_card_sum['U'] - 1)/ (1 + np.exp(-df_card_sum['t'])))
df_sum_m['U*']  = 1 + ((df_sum_m['U'] - 1)/ (1 + np.exp(-df_sum_m['t'])))

df_card_sum = df_card_sum.sort_values('U*', ascending = False)
df_card_sum = df_card_sum[df_card_sum['U*'] !=float("inf")]

df_sum_m = df_sum_m.sort_values('U*', ascending = False)
df_sum_m = df_sum_m[df_sum_m['U*'] !=float("inf")]

card_final = df_card_sum.head(40)
card_final.round(2).to_csv('card_benford.csv', index=False)

merc_final = df_sum_m.head(40)
merc_final.round(2).to_csv('merc_benford.csv', index=False)

merc_final
