==> 9-modeling-card.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
#from sklearn.linear_model import LogisticRegression
#from sklearn.neural_network import MLPClassifier


var_30 = pd.read_csv('top_30_wrapper.csv')
all_df = pd.read_csv('all_df_var.csv')

var_model = var_30['field'].to_list() + ['Fraud'] + ['Date']
pre_df = all_df[var_model]

train = pre_df[pre_df.Date <= '2010-08-31']
oot = pre_df[pre_df.Date > '2010-08-31']

FDR3 = pd.DataFrame([0, 1, 2, 3, 4, 5, 6, 7, 8, 9], columns=['train'])
FDR3['test'] = 0
FDR3['oot'] = 0
FDR3

var_model_final = var_30['field'].to_list()# + ['Fraud']
X = train[var_model_final]
y = train['Fraud'].values
X_oot = oot[var_model_final]
y_oot = oot['Fraud'].values

scaler = StandardScaler()    
X = scaler.fit_transform(X)
X_oot = scaler.fit_transform(X_oot)

### Logistic regression

evaluation = pd.DataFrame([1, 2, 3, 4, 5, 6, 7, 8, 9, 10], columns=['Logistic Regression'])
evaluation['max_iter'] = [15000, 15000, 10000, 10000, 5000, 5000, 2000, 2000, 2000, 1000]
evaluation['C'] = [1, 1, 2, 1, 1.5, 1.5, 1.5, 0.5, 0.5, 0.5]
evaluation['Class_weight'] = 0.0107
evaluation['solver'] = ['lbfgs', 'liblinear', 'lbfgs', 'liblinear', 'lbfgs', 'liblinear', 'lbfgs', 'liblinear', 'liblinear', 'lbfgs']
evaluation['train'] = 0
evaluation['test'] = 0
evaluation['oot'] = 0
evaluation

for idx in evaluation.index:
    max_iter_it = evaluation.loc[idx, 'max_iter']
    C_it = evaluation.loc[idx, 'C']
    weight_it = evaluation.loc[idx, 'Class_weight']
    solve_it = evaluation.loc[idx, 'solver']

    for niter in range(10):
        X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3)

        X_oot_eval = pd.DataFrame(X_oot.copy())
        X_train_eval = pd.DataFrame(X_train.copy())
        X_test_eval = pd.DataFrame(X_test.copy())

        clf = LogisticRegression(max_iter=max_iter_it, 
                                 C=C_it, 
                                 class_weight={1:y_train.mean()}, 
                                 solver=solve_it)
        clf.fit(X_train, y_train)

        predictions = clf.predict_proba(X_train)[:, 1]

        X_train_eval['pred'] = predictions
        X_train_eval['Fraud'] = y_train
        topRows = int(round(X_train_eval.shape[0]*0.03))
        temp = X_train_eval.sort_values('pred', ascending=False).head(topRows)
        needed = temp.loc[:, 'Fraud']
        FDR3.loc[niter, 'train'] = sum(needed)/sum(X_train_eval.loc[:, 'Fraud'])


        predictions = clf.predict_proba(X_test)[:, 1]

        X_test_eval['pred'] = predictions
        X_test_eval['Fraud'] = y_test
        topRows = int(round(X_test_eval.shape[0]*0.03))
        temp = X_test_eval.sort_values('pred', ascending=False).head(topRows)
        needed = temp.loc[:, 'Fraud']
        FDR3.loc[niter, 'test'] = sum(needed)/sum(X_test_eval.loc[:, 'Fraud'])

        predictions = clf.predict_proba(X_oot)[:, 1]

        X_oot_eval['pred'] = predictions
        X_oot_eval['Fraud'] = y_oot
        topRows = int(round(X_oot_eval.shape[0]*0.03))
        temp = X_oot_eval.sort_values('pred', ascending=False).head(topRows)
        needed = temp.loc[:, 'Fraud']
        FDR3.loc[niter, 'oot'] = sum(needed)/sum(X_oot_eval.loc[:, 'Fraud'])

        #print(niter, FDR3.loc[niter, 'train'], FDR3.loc[niter, 'test'], FDR3.loc[niter, 'oot'])
        
    evaluation.loc[idx, 'train'] = FDR3.mean()['train']
    evaluation.loc[idx, 'test'] = FDR3.mean()['test']
    evaluation.loc[idx, 'oot'] = FDR3.mean()['oot']

logit = evaluation.copy()
logit.to_csv('logit.csv', index=False)
logit

### MLP Classifier

nr_hnodes = range(1, 5)
hls = list(zip(nr_hnodes)) + list(zip(nr_hnodes, nr_hnodes))

evaluation = pd.DataFrame([1, 2, 3, 4, 5, 6, 7, 8], columns=['MLP Classifier'])
evaluation['max_iter'] = 10000
evaluation['alpha'] = [0.0001, 0.0001, 0.001, 0.001, 0.01, 0.01, 0.01, 0.05]
evaluation['activation']=['relu', 'tanh','relu', 'logistic', 'tanh', 'relu', 'tanh','relu']
evaluation['solver'] = ['lbfgs', 'lbfgs', 'lbfgs', 'lbfgs', 'adam', 'adam', 'adam', 'adam']
evaluation['hidden_layer_sizes'] = hls
evaluation['train'] = 0
evaluation['test'] = 0
evaluation['oot'] = 0
evaluation

for idx in evaluation.index:
    max_iter_it = evaluation.loc[idx, 'max_iter']
    alpha_it = evaluation.loc[idx, 'alpha']
    solve_it = evaluation.loc[idx, 'solver']
    act_it = evaluation.loc[idx, 'activation']
    hid_it = evaluation.loc[idx, 'hidden_layer_sizes']
    for niter in range(10):
        X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3)

        X_oot_eval = pd.DataFrame(X_oot.copy())
        X_train_eval = pd.DataFrame(X_train.copy())
        X_test_eval = pd.DataFrame(X_test.copy())


        mlp = MLPClassifier(activation=act_it,
                       solver=solve_it,
                       alpha=alpha_it,
                       hidden_layer_sizes=hid_it,
                       max_iter=max_iter_it,
                      )#.fit(X_train, y_train)
        mlp.fit(X_train, y_train)

        predictions = mlp.predict_proba(X_train)[:, 1]

        X_train_eval['pred'] = predictions
        X_train_eval['Fraud'] = y_train
        topRows = int(round(X_train_eval.shape[0]*0.03))
        temp = X_train_eval.sort_values('pred', ascending=False).head(topRows)
        needed = temp.loc[:, 'Fraud']
        FDR3.loc[niter, 'train'] = sum(needed)/sum(X_train_eval.loc[:, 'Fraud'])


        predictions = mlp.predict_proba(X_test)[:, 1]

        X_test_eval['pred'] = predictions
        X_test_eval['Fraud'] = y_test
        topRows = int(round(X_test_eval.shape[0]*0.03))
        temp = X_test_eval.sort_values('pred', ascending=False).head(topRows)
        needed = temp.loc[:, 'Fraud']
        FDR3.loc[niter, 'test'] = sum(needed)/sum(X_test_eval.loc[:, 'Fraud'])

        predictions = mlp.predict_proba(X_oot)[:, 1]

        X_oot_eval['pred'] = predictions
        X_oot_eval['Fraud'] = y_oot
        topRows = int(round(X_oot_eval.shape[0]*0.03))
        temp = X_oot_eval.sort_values('pred', ascending=False).head(topRows)
        needed = temp.loc[:, 'Fraud']
        FDR3.loc[niter, 'oot'] = sum(needed)/sum(X_oot_eval.loc[:, 'Fraud'])

        #print(niter, FDR3.loc[niter, 'train'], FDR3.loc[niter, 'test'], FDR3.loc[niter, 'oot'])
        
    evaluation.loc[idx, 'train'] = FDR3.mean()['train']
    evaluation.loc[idx, 'test'] = FDR3.mean()['test']
    evaluation.loc[idx, 'oot'] = FDR3.mean()['oot']
    print(idx)

mlp_result = evaluation.copy()
mlp_result.to_csv('mlp_result.csv', index=False)
mlp_result

### xgboost

evaluation = pd.DataFrame([1, 2, 3, 4, 5, 6, 7], columns=['xgboost'])
evaluation['objective'] = 'binary:logistic'
evaluation['colsample_bytree'] = [0.3, 0.3, 0.7, 0.7, 0.7, 1, 1]
evaluation['n_estimators']= [50, 100, 50, 100, 150, 100, 150]
evaluation['max_depth']=[2, 5, 7, 10, 2, 5, 7]
evaluation['eta'] = [0.01,0.1,0.5,0.9, 0.01, 0.5, 0.1]
evaluation['train'] = 0
evaluation['test'] = 0
evaluation['oot'] = 0
evaluation

import xgboost as xgb

for idx in evaluation.index:
    obj_it = evaluation.loc[idx, 'objective']
    col_it = evaluation.loc[idx, 'colsample_bytree']
    estim_it = evaluation.loc[idx, 'n_estimators']
    dep_it = evaluation.loc[idx, 'max_depth']
    eta_it = evaluation.loc[idx, 'eta']
    
    for niter in range(10):
        X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3)

        X_oot_eval = pd.DataFrame(X_oot.copy())
        X_train_eval = pd.DataFrame(X_train.copy())
        X_test_eval = pd.DataFrame(X_test.copy())


        gbm = xgb.XGBClassifier(use_label_encoder = False, 
                               objective=obj_it,
                               colsample_bytree=col_it, 
                               n_estimators=estim_it, 
                               max_depth=dep_it, 
                               eta=eta_it
                               )
    
        
        gbm.fit(X_train, y_train)

        predictions = gbm.predict_proba(X_train)[:, 1]

        X_train_eval['pred'] = predictions
        X_train_eval['Fraud'] = y_train
        topRows = int(round(X_train_eval.shape[0]*0.03))
        temp = X_train_eval.sort_values('pred', ascending=False).head(topRows)
        needed = temp.loc[:, 'Fraud']
        FDR3.loc[niter, 'train'] = sum(needed)/sum(X_train_eval.loc[:, 'Fraud'])


        predictions = gbm.predict_proba(X_test)[:, 1]

        X_test_eval['pred'] = predictions
        X_test_eval['Fraud'] = y_test
        topRows = int(round(X_test_eval.shape[0]*0.03))
        temp = X_test_eval.sort_values('pred', ascending=False).head(topRows)
        needed = temp.loc[:, 'Fraud']
        FDR3.loc[niter, 'test'] = sum(needed)/sum(X_test_eval.loc[:, 'Fraud'])

        predictions = gbm.predict_proba(X_oot)[:, 1]

        X_oot_eval['pred'] = predictions
        X_oot_eval['Fraud'] = y_oot
        topRows = int(round(X_oot_eval.shape[0]*0.03))
        temp = X_oot_eval.sort_values('pred', ascending=False).head(topRows)
        needed = temp.loc[:, 'Fraud']
        FDR3.loc[niter, 'oot'] = sum(needed)/sum(X_oot_eval.loc[:, 'Fraud'])

        #print(niter, FDR3.loc[niter, 'train'], FDR3.loc[niter, 'test'], FDR3.loc[niter, 'oot'])
    print(idx)    
    evaluation.loc[idx, 'train'] = FDR3.mean()['train']
    evaluation.loc[idx, 'test'] = FDR3.mean()['test']
    evaluation.loc[idx, 'oot'] = FDR3.mean()['oot']
    print(idx)

xgb_result = evaluation.copy()
xgb_result.to_csv('xgb_result.csv', index=False)
xgb_result

### random forest

evaluation = pd.DataFrame([1, 2, 3, 4, 5, 6], columns=['ramdom forest'])

evaluation['max_features'] = ['auto', 'sqrt', 'auto', 'sqrt', 'auto','sqrt']
evaluation['n_estimators']= [50, 100, 50, 100, 100, 100]
evaluation['max_depth']=[2, 5, 7, 10, 5, 7]
evaluation['min_samples_split'] = [2, 5, 2, 5, 3, 5]
evaluation['min_samples_leaf'] = [1, 2, 4, 1, 2, 4]
evaluation['bootstrap'] = [True, False, True, True, True, False]

evaluation['train'] = 0
evaluation['test'] = 0
evaluation['oot'] = 0
evaluation

from sklearn.ensemble import RandomForestClassifier

for idx in evaluation.index:
    maxfet_it = evaluation.loc[idx, 'max_features']
    samsplt_it = evaluation.loc[idx, 'min_samples_split']
    leaf_it = evaluation.loc[idx, 'min_samples_leaf']
    estim_it = evaluation.loc[idx, 'n_estimators']
    dep_it = evaluation.loc[idx, 'max_depth']
    boot_it = evaluation.loc[idx, 'bootstrap']
    
    for niter in range(10):
        X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3)

        X_oot_eval = pd.DataFrame(X_oot.copy())
        X_train_eval = pd.DataFrame(X_train.copy())
        X_test_eval = pd.DataFrame(X_test.copy())
        
        rfc = RandomForestClassifier(n_estimators=estim_it, 
                                    max_features = maxfet_it, 
                                    min_samples_split=samsplt_it, 
                                    min_samples_leaf=leaf_it, 
                                    max_depth=dep_it, 
                                    bootstrap=boot_it
                                    )

    
        
        rfc.fit(X_train, y_train)

        predictions = rfc.predict_proba(X_train)[:, 1]

        X_train_eval['pred'] = predictions
        X_train_eval['Fraud'] = y_train
        topRows = int(round(X_train_eval.shape[0]*0.03))
        temp = X_train_eval.sort_values('pred', ascending=False).head(topRows)
        needed = temp.loc[:, 'Fraud']
        FDR3.loc[niter, 'train'] = sum(needed)/sum(X_train_eval.loc[:, 'Fraud'])


        predictions = rfc.predict_proba(X_test)[:, 1]

        X_test_eval['pred'] = predictions
        X_test_eval['Fraud'] = y_test
        topRows = int(round(X_test_eval.shape[0]*0.03))
        temp = X_test_eval.sort_values('pred', ascending=False).head(topRows)
        needed = temp.loc[:, 'Fraud']
        FDR3.loc[niter, 'test'] = sum(needed)/sum(X_test_eval.loc[:, 'Fraud'])

        predictions = rfc.predict_proba(X_oot)[:, 1]

        X_oot_eval['pred'] = predictions
        X_oot_eval['Fraud'] = y_oot
        topRows = int(round(X_oot_eval.shape[0]*0.03))
        temp = X_oot_eval.sort_values('pred', ascending=False).head(topRows)
        needed = temp.loc[:, 'Fraud']
        FDR3.loc[niter, 'oot'] = sum(needed)/sum(X_oot_eval.loc[:, 'Fraud'])

        #print(niter, FDR3.loc[niter, 'train'], FDR3.loc[niter, 'test'], FDR3.loc[niter, 'oot'])
        
    evaluation.loc[idx, 'train'] = FDR3.mean()['train']
    evaluation.loc[idx, 'test'] = FDR3.mean()['test']
    evaluation.loc[idx, 'oot'] = FDR3.mean()['oot']
    print(idx)

rfm_result = evaluation.copy()
rfm_result.to_csv('rfm_result.csv', index=False)
rfm_result
